requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
requests